Added verification for new clients connecting
All connections now send all messages using JSON. A message.py file contains all the classes for formatting the dictionaries which will be converted into JSON for consistency. New connections now need to send a verification message containing a username and password. If the first message a connection sends isn't in the proper format for a verification request message, the connection is closed. If the username or password is incorrect in the verification request message, the connection is closed.